Use intermediate files in both batch creation and applying processes if necessary
On Windows, xdelta won't understand any characters outside some archaic codepage even if
the codepage is set to UTF-8 (65001) beforehand. To be able to still handle files with such
characters in their names, the files are first copied to a second location with a safe
filename, then processed, after which the intermediate files are deleted.